refactor: review指摘箇所修正
