+++ scripts/generateHolidays.ts
@@ -17,7 +17,7 @@ const holidays = raw.map((h) => ({
 }))
 
 // Keep stable order
-holidays.sort((a, b) => (a.date < b.date ? -1 : a.date > b.date ? 1 : 0))
+holidays.sort((a, b) => a.date.localeCompare(b.date))
 
 const out = {
   year,
